+++ lib/schleuder/cli.rb
@@ -227,7 +227,9 @@ module Schleuder
           sub.save!
         else
           adminfpr = member['key_fingerprint'] || list.keys(member['email']).first.fingerprint
-          list.subscribe(member['email'], adminfpr, true)
+          # if we didn't find an already imported  subscription for the admin
+          # address, it wasn't a member, so we don't enable delivery for it
+          list.subscribe(member['email'], adminfpr, true, false)
         end
       end